+++ tests/test_cluster_has_leader.py
@@ -2,12 +2,12 @@ from click.testing import CliRunner
 
 from check_patroni.cli import main
 
+from . import PatroniAPI
 
-def test_cluster_has_leader_ok(runner: CliRunner, fake_restapi) -> None:
-    fake_restapi("cluster_has_leader_ok")
-    result = runner.invoke(
-        main, ["-e", "https://10.20.199.3:8008", "cluster_has_leader"]
-    )
+
+def test_cluster_has_leader_ok(runner: CliRunner, patroni_api: PatroniAPI) -> None:
+    with patroni_api.routes({"cluster": "cluster_has_leader_ok.json"}):
+        result = runner.invoke(main, ["-e", patroni_api.endpoint, "cluster_has_leader"])
     assert result.exit_code == 0
     assert (
         result.stdout
@@ -15,11 +15,11 @@ def test_cluster_has_leader_ok(runner: CliRunner, fake_restapi) -> None:
     )
 
 
-def test_cluster_has_leader_ok_standby_leader(runner: CliRunner, fake_restapi) -> None:
-    fake_restapi("cluster_has_leader_ok_standby_leader")
-    result = runner.invoke(
-        main, ["-e", "https://10.20.199.3:8008", "cluster_has_leader"]
-    )
+def test_cluster_has_leader_ok_standby_leader(
+    runner: CliRunner, patroni_api: PatroniAPI
+) -> None:
+    with patroni_api.routes({"cluster": "cluster_has_leader_ok_standby_leader.json"}):
+        result = runner.invoke(main, ["-e", patroni_api.endpoint, "cluster_has_leader"])
     assert result.exit_code == 0
     assert (
         result.stdout
@@ -27,11 +27,9 @@ def test_cluster_has_leader_ok_standby_leader(runner: CliRunner, fake_restapi) -
     )
 
 
-def test_cluster_has_leader_ko(runner: CliRunner, fake_restapi) -> None:
-    fake_restapi("cluster_has_leader_ko")
-    result = runner.invoke(
-        main, ["-e", "https://10.20.199.3:8008", "cluster_has_leader"]
-    )
+def test_cluster_has_leader_ko(runner: CliRunner, patroni_api: PatroniAPI) -> None:
+    with patroni_api.routes({"cluster": "cluster_has_leader_ko.json"}):
+        result = runner.invoke(main, ["-e", patroni_api.endpoint, "cluster_has_leader"])
     assert result.exit_code == 2
     assert (
         result.stdout