+++ tests/test_cluster_has_replica.py
@@ -1,16 +1,29 @@
+from pathlib import Path
+from typing import Iterator, Union
+
+import pytest
 from click.testing import CliRunner
 
 from check_patroni.cli import main
 
+from . import PatroniAPI, cluster_api_set_replica_running
+
+
+@pytest.fixture
+def cluster_has_replica_ok(
+    patroni_api: PatroniAPI, old_replica_state: bool, datadir: Path, tmp_path: Path
+) -> Iterator[None]:
+    path: Union[str, Path] = "cluster_has_replica_ok.json"
+    if old_replica_state:
+        path = cluster_api_set_replica_running(datadir / path, tmp_path)
+    with patroni_api.routes({"cluster": path}):
+        yield None
+
 
 # TODO Lag threshold tests
-def test_cluster_has_relica_ok(
-    runner: CliRunner, fake_restapi, old_replica_state: bool
-) -> None:
-    fake_restapi("cluster_has_replica_ok", use_old_replica_state=old_replica_state)
-    result = runner.invoke(
-        main, ["-e", "https://10.20.199.3:8008", "cluster_has_replica"]
-    )
+@pytest.mark.usefixtures("cluster_has_replica_ok")
+def test_cluster_has_relica_ok(runner: CliRunner, patroni_api: PatroniAPI) -> None:
+    result = runner.invoke(main, ["-e", patroni_api.endpoint, "cluster_has_replica"])
     assert result.exit_code == 0
     assert (
         result.stdout
@@ -18,15 +31,15 @@ def test_cluster_has_relica_ok(
     )
 
 
+@pytest.mark.usefixtures("cluster_has_replica_ok")
 def test_cluster_has_replica_ok_with_count_thresholds(
-    runner: CliRunner, fake_restapi, old_replica_state: bool
+    runner: CliRunner, patroni_api: PatroniAPI
 ) -> None:
-    fake_restapi("cluster_has_replica_ok", use_old_replica_state=old_replica_state)
     result = runner.invoke(
         main,
         [
             "-e",
-            "https://10.20.199.3:8008",
+            patroni_api.endpoint,
             "cluster_has_replica",
             "--warning",
             "@1",
@@ -41,15 +54,15 @@ def test_cluster_has_replica_ok_with_count_thresholds(
     )
 
 
+@pytest.mark.usefixtures("cluster_has_replica_ok")
 def test_cluster_has_replica_ok_with_sync_count_thresholds(
-    runner: CliRunner, fake_restapi, old_replica_state: bool
+    runner: CliRunner, patroni_api: PatroniAPI
 ) -> None:
-    fake_restapi("cluster_has_replica_ok", use_old_replica_state=old_replica_state)
     result = runner.invoke(
         main,
         [
             "-e",
-            "https://10.20.199.3:8008",
+            patroni_api.endpoint,
             "cluster_has_replica",
             "--sync-warning",
             "1:",
@@ -62,15 +75,26 @@ def test_cluster_has_replica_ok_with_sync_count_thresholds(
     )
 
 
+@pytest.fixture
+def cluster_has_replica_ok_lag(
+    patroni_api: PatroniAPI, datadir: Path, tmp_path: Path, old_replica_state: bool
+) -> Iterator[None]:
+    path: Union[str, Path] = "cluster_has_replica_ok_lag.json"
+    if old_replica_state:
+        path = cluster_api_set_replica_running(datadir / path, tmp_path)
+    with patroni_api.routes({"cluster": path}):
+        yield None
+
+
+@pytest.mark.usefixtures("cluster_has_replica_ok_lag")
 def test_cluster_has_replica_ok_with_count_thresholds_lag(
-    runner: CliRunner, fake_restapi, old_replica_state: bool
+    runner: CliRunner, patroni_api: PatroniAPI
 ) -> None:
-    fake_restapi("cluster_has_replica_ok_lag", use_old_replica_state=old_replica_state)
     result = runner.invoke(
         main,
         [
             "-e",
-            "https://10.20.199.3:8008",
+            patroni_api.endpoint,
             "cluster_has_replica",
             "--warning",
             "@1",
@@ -87,15 +111,26 @@ def test_cluster_has_replica_ok_with_count_thresholds_lag(
     )
 
 
+@pytest.fixture
+def cluster_has_replica_ko(
+    patroni_api: PatroniAPI, old_replica_state: bool, datadir: Path, tmp_path: Path
+) -> Iterator[None]:
+    path: Union[str, Path] = "cluster_has_replica_ko.json"
+    if old_replica_state:
+        path = cluster_api_set_replica_running(datadir / path, tmp_path)
+    with patroni_api.routes({"cluster": path}):
+        yield None
+
+
+@pytest.mark.usefixtures("cluster_has_replica_ko")
 def test_cluster_has_replica_ko_with_count_thresholds(
-    runner: CliRunner, fake_restapi, old_replica_state: bool
+    runner: CliRunner, patroni_api: PatroniAPI
 ) -> None:
-    fake_restapi("cluster_has_replica_ko", use_old_replica_state=old_replica_state)
     result = runner.invoke(
         main,
         [
             "-e",
-            "https://10.20.199.3:8008",
+            patroni_api.endpoint,
             "cluster_has_replica",
             "--warning",
             "@1",
@@ -110,15 +145,15 @@ def test_cluster_has_replica_ko_with_count_thresholds(
     )
 
 
+@pytest.mark.usefixtures("cluster_has_replica_ko")
 def test_cluster_has_replica_ko_with_sync_count_thresholds(
-    runner: CliRunner, fake_restapi, old_replica_state: bool
+    runner: CliRunner, patroni_api: PatroniAPI
 ) -> None:
-    fake_restapi("cluster_has_replica_ko", use_old_replica_state=old_replica_state)
     result = runner.invoke(
         main,
         [
             "-e",
-            "https://10.20.199.3:8008",
+            patroni_api.endpoint,
             "cluster_has_replica",
             "--sync-warning",
             "2:",
@@ -133,15 +168,26 @@ def test_cluster_has_replica_ko_with_sync_count_thresholds(
     )
 
 
+@pytest.fixture
+def cluster_has_replica_ko_lag(
+    patroni_api: PatroniAPI, old_replica_state: bool, datadir: Path, tmp_path: Path
+) -> Iterator[None]:
+    path: Union[str, Path] = "cluster_has_replica_ko_lag.json"
+    if old_replica_state:
+        path = cluster_api_set_replica_running(datadir / path, tmp_path)
+    with patroni_api.routes({"cluster": path}):
+        yield None
+
+
+@pytest.mark.usefixtures("cluster_has_replica_ko_lag")
 def test_cluster_has_replica_ko_with_count_thresholds_and_lag(
-    runner: CliRunner, fake_restapi, old_replica_state: bool
+    runner: CliRunner, patroni_api: PatroniAPI
 ) -> None:
-    fake_restapi("cluster_has_replica_ko_lag", use_old_replica_state=old_replica_state)
     result = runner.invoke(
         main,
         [
             "-e",
-            "https://10.20.199.3:8008",
+            patroni_api.endpoint,
             "cluster_has_replica",
             "--warning",
             "@1",