+++ tests/test_node_is_alive.py
@@ -1,11 +1,19 @@
+from pathlib import Path
+
 from click.testing import CliRunner
 
 from check_patroni.cli import main
 
+from . import PatroniAPI
 
-def test_node_is_alive_ok(runner: CliRunner, fake_restapi) -> None:
-    fake_restapi(None)
-    result = runner.invoke(main, ["-e", "https://10.20.199.3:8008", "node_is_alive"])
+
+def test_node_is_alive_ok(
+    runner: CliRunner, patroni_api: PatroniAPI, tmp_path: Path
+) -> None:
+    liveness = tmp_path / "liveness"
+    liveness.touch()
+    with patroni_api.routes({"liveness": liveness}):
+        result = runner.invoke(main, ["-e", patroni_api.endpoint, "node_is_alive"])
     assert result.exit_code == 0
     assert (
         result.stdout
@@ -13,9 +21,8 @@ def test_node_is_alive_ok(runner: CliRunner, fake_restapi) -> None:
     )
 
 
-def test_node_is_alive_ko(runner: CliRunner, fake_restapi) -> None:
-    fake_restapi(None, status=404)
-    result = runner.invoke(main, ["-e", "https://10.20.199.3:8008", "node_is_alive"])
+def test_node_is_alive_ko(runner: CliRunner, patroni_api: PatroniAPI) -> None:
+    result = runner.invoke(main, ["-e", patroni_api.endpoint, "node_is_alive"])
     assert result.exit_code == 2
     assert (
         result.stdout