+++ tests/test_node_is_pending_restart.py
@@ -2,12 +2,14 @@ from click.testing import CliRunner
 
 from check_patroni.cli import main
 
+from . import PatroniAPI
 
-def test_node_is_pending_restart_ok(runner: CliRunner, fake_restapi) -> None:
-    fake_restapi("node_is_pending_restart_ok")
-    result = runner.invoke(
-        main, ["-e", "https://10.20.199.3:8008", "node_is_pending_restart"]
-    )
+
+def test_node_is_pending_restart_ok(runner: CliRunner, patroni_api: PatroniAPI) -> None:
+    with patroni_api.routes({"patroni": "node_is_pending_restart_ok.json"}):
+        result = runner.invoke(
+            main, ["-e", patroni_api.endpoint, "node_is_pending_restart"]
+        )
     assert result.exit_code == 0
     assert (
         result.stdout
@@ -15,11 +17,11 @@ def test_node_is_pending_restart_ok(runner: CliRunner, fake_restapi) -> None:
     )
 
 
-def test_node_is_pending_restart_ko(runner: CliRunner, fake_restapi) -> None:
-    fake_restapi("node_is_pending_restart_ko")
-    result = runner.invoke(
-        main, ["-e", "https://10.20.199.3:8008", "node_is_pending_restart"]
-    )
+def test_node_is_pending_restart_ko(runner: CliRunner, patroni_api: PatroniAPI) -> None:
+    with patroni_api.routes({"patroni": "node_is_pending_restart_ko.json"}):
+        result = runner.invoke(
+            main, ["-e", patroni_api.endpoint, "node_is_pending_restart"]
+        )
     assert result.exit_code == 2
     assert (
         result.stdout