+++ tests/test_node_is_primary.py
@@ -2,10 +2,12 @@ from click.testing import CliRunner
 
 from check_patroni.cli import main
 
+from . import PatroniAPI
 
-def test_node_is_primary_ok(runner: CliRunner, fake_restapi) -> None:
-    fake_restapi("node_is_primary_ok")
-    result = runner.invoke(main, ["-e", "https://10.20.199.3:8008", "node_is_primary"])
+
+def test_node_is_primary_ok(runner: CliRunner, patroni_api: PatroniAPI) -> None:
+    with patroni_api.routes({"primary": "node_is_primary_ok.json"}):
+        result = runner.invoke(main, ["-e", patroni_api.endpoint, "node_is_primary"])
     assert result.exit_code == 0
     assert (
         result.stdout
@@ -13,9 +15,8 @@ def test_node_is_primary_ok(runner: CliRunner, fake_restapi) -> None:
     )
 
 
-def test_node_is_primary_ko(runner: CliRunner, fake_restapi) -> None:
-    fake_restapi("node_is_primary_ko", status=503)
-    result = runner.invoke(main, ["-e", "https://10.20.199.3:8008", "node_is_primary"])
+def test_node_is_primary_ko(runner: CliRunner, patroni_api: PatroniAPI) -> None:
+    result = runner.invoke(main, ["-e", patroni_api.endpoint, "node_is_primary"])
     assert result.exit_code == 2
     assert (
         result.stdout