+++ tests/test_node_is_replica.py
@@ -1,11 +1,27 @@
+from typing import Iterator
+
+import pytest
 from click.testing import CliRunner
 
 from check_patroni.cli import main
 
+from . import PatroniAPI
 
-def test_node_is_replica_ok(runner: CliRunner, fake_restapi) -> None:
-    fake_restapi("node_is_replica_ok")
-    result = runner.invoke(main, ["-e", "https://10.20.199.3:8008", "node_is_replica"])
+
+@pytest.fixture
+def node_is_replica_ok(patroni_api: PatroniAPI) -> Iterator[None]:
+    with patroni_api.routes(
+        {
+            k: "node_is_replica_ok.json"
+            for k in ("replica", "synchronous", "asynchronous")
+        }
+    ):
+        yield None
+
+
+@pytest.mark.usefixtures("node_is_replica_ok")
+def test_node_is_replica_ok(runner: CliRunner, patroni_api: PatroniAPI) -> None:
+    result = runner.invoke(main, ["-e", patroni_api.endpoint, "node_is_replica"])
     assert result.exit_code == 0
     assert (
         result.stdout
@@ -13,9 +29,8 @@ def test_node_is_replica_ok(runner: CliRunner, fake_restapi) -> None:
     )
 
 
-def test_node_is_replica_ko(runner: CliRunner, fake_restapi) -> None:
-    fake_restapi("node_is_replica_ko", status=503)
-    result = runner.invoke(main, ["-e", "https://10.20.199.3:8008", "node_is_replica"])
+def test_node_is_replica_ko(runner: CliRunner, patroni_api: PatroniAPI) -> None:
+    result = runner.invoke(main, ["-e", patroni_api.endpoint, "node_is_replica"])
     assert result.exit_code == 2
     assert (
         result.stdout
@@ -23,11 +38,10 @@ def test_node_is_replica_ko(runner: CliRunner, fake_restapi) -> None:
     )
 
 
-def test_node_is_replica_ko_lag(runner: CliRunner, fake_restapi) -> None:
+def test_node_is_replica_ko_lag(runner: CliRunner, patroni_api: PatroniAPI) -> None:
     # We don't do the check ourselves, patroni does it and changes the return code
-    fake_restapi("node_is_replica_ok", status=503)
     result = runner.invoke(
-        main, ["-e", "https://10.20.199.3:8008", "node_is_replica", "--max-lag", "100"]
+        main, ["-e", patroni_api.endpoint, "node_is_replica", "--max-lag", "100"]
     )
     assert result.exit_code == 2
     assert (
@@ -35,12 +49,11 @@ def test_node_is_replica_ko_lag(runner: CliRunner, fake_restapi) -> None:
         == "NODEISREPLICA CRITICAL - This node is not a running replica with no noloadbalance tag and a lag under 100. | is_replica=0;;@0\n"
     )
 
-    fake_restapi("node_is_replica_ok", status=503)
     result = runner.invoke(
         main,
         [
             "-e",
-            "https://10.20.199.3:8008",
+            patroni_api.endpoint,
             "node_is_replica",
             "--is-async",
             "--max-lag",
@@ -54,11 +67,11 @@ def test_node_is_replica_ko_lag(runner: CliRunner, fake_restapi) -> None:
     )
 
 
-def test_node_is_replica_sync_ok(runner: CliRunner, fake_restapi) -> None:
+@pytest.mark.usefixtures("node_is_replica_ok")
+def test_node_is_replica_sync_ok(runner: CliRunner, patroni_api: PatroniAPI) -> None:
     # We don't do the check ourselves, patroni does it and changes the return code
-    fake_restapi("node_is_replica_ok")
     result = runner.invoke(
-        main, ["-e", "https://10.20.199.3:8008", "node_is_replica", "--is-sync"]
+        main, ["-e", patroni_api.endpoint, "node_is_replica", "--is-sync"]
     )
     assert result.exit_code == 0
     assert (
@@ -67,11 +80,10 @@ def test_node_is_replica_sync_ok(runner: CliRunner, fake_restapi) -> None:
     )
 
 
-def test_node_is_replica_sync_ko(runner: CliRunner, fake_restapi) -> None:
+def test_node_is_replica_sync_ko(runner: CliRunner, patroni_api: PatroniAPI) -> None:
     # We don't do the check ourselves, patroni does it and changes the return code
-    fake_restapi("node_is_replica_ok", status=503)
     result = runner.invoke(
-        main, ["-e", "https://10.20.199.3:8008", "node_is_replica", "--is-sync"]
+        main, ["-e", patroni_api.endpoint, "node_is_replica", "--is-sync"]
     )
     assert result.exit_code == 2
     assert (
@@ -80,11 +92,11 @@ def test_node_is_replica_sync_ko(runner: CliRunner, fake_restapi) -> None:
     )
 
 
-def test_node_is_replica_async_ok(runner: CliRunner, fake_restapi) -> None:
+@pytest.mark.usefixtures("node_is_replica_ok")
+def test_node_is_replica_async_ok(runner: CliRunner, patroni_api: PatroniAPI) -> None:
     # We don't do the check ourselves, patroni does it and changes the return code
-    fake_restapi("node_is_replica_ok")
     result = runner.invoke(
-        main, ["-e", "https://10.20.199.3:8008", "node_is_replica", "--is-async"]
+        main, ["-e", patroni_api.endpoint, "node_is_replica", "--is-async"]
     )
     assert result.exit_code == 0
     assert (
@@ -93,11 +105,10 @@ def test_node_is_replica_async_ok(runner: CliRunner, fake_restapi) -> None:
     )
 
 
-def test_node_is_replica_async_ko(runner: CliRunner, fake_restapi) -> None:
+def test_node_is_replica_async_ko(runner: CliRunner, patroni_api: PatroniAPI) -> None:
     # We don't do the check ourselves, patroni does it and changes the return code
-    fake_restapi("node_is_replica_ok", status=503)
     result = runner.invoke(
-        main, ["-e", "https://10.20.199.3:8008", "node_is_replica", "--is-async"]
+        main, ["-e", patroni_api.endpoint, "node_is_replica", "--is-async"]
     )
     assert result.exit_code == 2
     assert (
@@ -106,14 +117,14 @@ def test_node_is_replica_async_ko(runner: CliRunner, fake_restapi) -> None:
     )
 
 
-def test_node_is_replica_params(runner: CliRunner, fake_restapi) -> None:
+@pytest.mark.usefixtures("node_is_replica_ok")
+def test_node_is_replica_params(runner: CliRunner, patroni_api: PatroniAPI) -> None:
     # We don't do the check ourselves, patroni does it and changes the return code
-    fake_restapi("node_is_replica_ok")
     result = runner.invoke(
         main,
         [
             "-e",
-            "https://10.20.199.3:8008",
+            patroni_api.endpoint,
             "node_is_replica",
             "--is-async",
             "--is-sync",
@@ -126,12 +137,11 @@ def test_node_is_replica_params(runner: CliRunner, fake_restapi) -> None:
     )
 
     # We don't do the check ourselves, patroni does it and changes the return code
-    fake_restapi("node_is_replica_ok")
     result = runner.invoke(
         main,
         [
             "-e",
-            "https://10.20.199.3:8008",
+            patroni_api.endpoint,
             "node_is_replica",
             "--is-sync",
             "--max-lag",