+++ tests/test_node_patroni_version.py
@@ -1,15 +1,25 @@
+from typing import Iterator
+
+import pytest
 from click.testing import CliRunner
 
 from check_patroni.cli import main
 
+from . import PatroniAPI
 
-def test_node_patroni_version_ok(runner: CliRunner, fake_restapi) -> None:
-    fake_restapi("node_patroni_version")
+
+@pytest.fixture(scope="module", autouse=True)
+def node_patroni_version(patroni_api: PatroniAPI) -> Iterator[None]:
+    with patroni_api.routes({"patroni": "node_patroni_version.json"}):
+        yield None
+
+
+def test_node_patroni_version_ok(runner: CliRunner, patroni_api: PatroniAPI) -> None:
     result = runner.invoke(
         main,
         [
             "-e",
-            "https://10.20.199.3:8008",
+            patroni_api.endpoint,
             "node_patroni_version",
             "--patroni-version",
             "2.0.2",
@@ -22,13 +32,12 @@ def test_node_patroni_version_ok(runner: CliRunner, fake_restapi) -> None:
     )
 
 
-def test_node_patroni_version_ko(runner: CliRunner, fake_restapi) -> None:
-    fake_restapi("node_patroni_version")
+def test_node_patroni_version_ko(runner: CliRunner, patroni_api: PatroniAPI) -> None:
     result = runner.invoke(
         main,
         [
             "-e",
-            "https://10.20.199.3:8008",
+            patroni_api.endpoint,
             "node_patroni_version",
             "--patroni-version",
             "1.0.0",