+++ tests/test_node_tl_has_changed.py
@@ -1,18 +1,30 @@
 from pathlib import Path
+from typing import Iterator
 
 import nagiosplugin
+import pytest
 from click.testing import CliRunner
 
 from check_patroni.cli import main
 
+from . import PatroniAPI
 
-def test_node_tl_has_changed_ok_with_timeline(runner: CliRunner, fake_restapi) -> None:
-    fake_restapi("node_tl_has_changed")
+
+@pytest.fixture
+def node_tl_has_changed(patroni_api: PatroniAPI) -> Iterator[None]:
+    with patroni_api.routes({"patroni": "node_tl_has_changed.json"}):
+        yield None
+
+
+@pytest.mark.usefixtures("node_tl_has_changed")
+def test_node_tl_has_changed_ok_with_timeline(
+    runner: CliRunner, patroni_api: PatroniAPI
+) -> None:
     result = runner.invoke(
         main,
         [
             "-e",
-            "https://10.20.199.3:8008",
+            patroni_api.endpoint,
             "node_tl_has_changed",
             "--timeline",
             "58",
@@ -25,19 +37,19 @@ def test_node_tl_has_changed_ok_with_timeline(runner: CliRunner, fake_restapi) -
     )
 
 
+@pytest.mark.usefixtures("node_tl_has_changed")
 def test_node_tl_has_changed_ok_with_state_file(
-    runner: CliRunner, fake_restapi, tmp_path: Path
+    runner: CliRunner, patroni_api: PatroniAPI, tmp_path: Path
 ) -> None:
     state_file = tmp_path / "node_tl_has_changed.state_file"
     with state_file.open("w") as f:
         f.write('{"timeline": 58}')
 
-    fake_restapi("node_tl_has_changed")
     result = runner.invoke(
         main,
         [
             "-e",
-            "https://10.20.199.3:8008",
+            patroni_api.endpoint,
             "node_tl_has_changed",
             "--state-file",
             str(state_file),
@@ -50,13 +62,15 @@ def test_node_tl_has_changed_ok_with_state_file(
     )
 
 
-def test_node_tl_has_changed_ko_with_timeline(runner: CliRunner, fake_restapi) -> None:
-    fake_restapi("node_tl_has_changed")
+@pytest.mark.usefixtures("node_tl_has_changed")
+def test_node_tl_has_changed_ko_with_timeline(
+    runner: CliRunner, patroni_api: PatroniAPI
+) -> None:
     result = runner.invoke(
         main,
         [
             "-e",
-            "https://10.20.199.3:8008",
+            patroni_api.endpoint,
             "node_tl_has_changed",
             "--timeline",
             "700",
@@ -69,20 +83,20 @@ def test_node_tl_has_changed_ko_with_timeline(runner: CliRunner, fake_restapi) -
     )
 
 
+@pytest.mark.usefixtures("node_tl_has_changed")
 def test_node_tl_has_changed_ko_with_state_file_and_save(
-    runner: CliRunner, fake_restapi, tmp_path: Path
+    runner: CliRunner, patroni_api: PatroniAPI, tmp_path: Path
 ) -> None:
     state_file = tmp_path / "node_tl_has_changed.state_file"
     with state_file.open("w") as f:
         f.write('{"timeline": 700}')
 
-    fake_restapi("node_tl_has_changed")
     # test without saving the new tl
     result = runner.invoke(
         main,
         [
             "-e",
-            "https://10.20.199.3:8008",
+            patroni_api.endpoint,
             "node_tl_has_changed",
             "--state-file",
             str(state_file),
@@ -106,7 +120,7 @@ def test_node_tl_has_changed_ko_with_state_file_and_save(
         main,
         [
             "-e",
-            "https://10.20.199.3:8008",
+            patroni_api.endpoint,
             "node_tl_has_changed",
             "--state-file",
             str(state_file),
@@ -127,17 +141,17 @@ def test_node_tl_has_changed_ko_with_state_file_and_save(
     assert new_tl == 58
 
 
+@pytest.mark.usefixtures("node_tl_has_changed")
 def test_node_tl_has_changed_params(
-    runner: CliRunner, fake_restapi, tmp_path: Path
+    runner: CliRunner, patroni_api: PatroniAPI, tmp_path: Path
 ) -> None:
     # This one is placed last because it seems like the exceptions are not flushed from stderr for the next tests.
     fake_state_file = tmp_path / "fake_file_name.state_file"
-    fake_restapi("node_tl_has_changed")
     result = runner.invoke(
         main,
         [
             "-e",
-            "https://10.20.199.3:8008",
+            patroni_api.endpoint,
             "node_tl_has_changed",
             "--timeline",
             "58",
@@ -151,9 +165,7 @@ def test_node_tl_has_changed_params(
         == "NODETLHASCHANGED UNKNOWN: click.exceptions.UsageError: Either --timeline or --state-file should be provided for this service\n"
     )
 
-    result = runner.invoke(
-        main, ["-e", "https://10.20.199.3:8008", "node_tl_has_changed"]
-    )
+    result = runner.invoke(main, ["-e", patroni_api.endpoint, "node_tl_has_changed"])
     assert result.exit_code == 3
     assert (
         result.stdout